Idleness for agreements stream and Flink bug workaround
Connected streams did not produce expected results, because agreements stream is idle. When connecting two streams, resulting watermark is the min of two watermarks - it always stayed at the event time of the agreement.

The workaround is to generate and send to Kafka an agreement in far future. This is obviously not production ready workaround, it applies to tests and only assuming that agreements are generated once, and then only incoming data records are generated.

Bug in Flink: https://issues.apache.org/jira/browse/FLINK-18934